+++ app/Repos/HackerNews/BookmarkedStories.php
@@ -12,19 +12,14 @@ class BookmarkedStories
     /**
      * Shows the list of bookmarked stories for the given user
      *
-     * @param int      $limit The maximum number of stories to show
-     * @param int      $offset The offset of list of stories
      * @param int|null $userId The identifier of the user
      *
-     * @return object Returns an object contained the total number of stories
-     *                and the stories.
+     * @return array Returns an array containing the ids of the bookmarked
+     *               stories
      */
-    public function bookmarkedStories(
-        int $limit = 20,
-        int $offset = 0,
-        ?int $userId = null
-    ) {
-        $hnIds = (new HackerNewsItemsBookmark())
+    public function bookmarkedStories(?int $userId = null): array
+    {
+        return (new HackerNewsItemsBookmark())
             ->select('hacker_news_item_id')
             ->where('user_id', $userId)
             ->orderBy('created_at', 'DESC')
@@ -32,8 +27,6 @@ class BookmarkedStories
             ->pluck('hacker_news_item_id')
             ->toArray()
         ;
-
-        return Utils::storiesListFromDb($hnIds, $limit, $offset);
     }
 
     /**
@@ -78,21 +71,23 @@ class BookmarkedStories
      * @param int $hackerNewsItemId The identifier of the story.
      * @param int $userId           The identifier of the user.
      *
-     * @return bool Returns true on success, false otherwise.
+     * @return int Returns the id of the destroyed bookmark, null otherwise
      */
     public function destroyBookmarkedStory(
         int $hackerNewsItemId,
         int $userId
-    ): bool {
+    ): ?int {
         $hackerNewsItemsBookmark = (new HackerNewsItemsBookmark())
             ->where('user_id', $userId)
             ->where('hacker_news_item_id', $hackerNewsItemId)
             ->first();
-        $isDeleted = false;
+        $id = null;
         if ($hackerNewsItemsBookmark) {
+            $id = $hackerNewsItemsBookmark->id;
             try {
-                $isDeleted = $hackerNewsItemsBookmark->delete();
+                $hackerNewsItemsBookmark->delete();
             } catch (Exception $exception) {
+                $id = null;
                 Log::error(
                     'Failed to destroy bookmark',
                     [
@@ -104,6 +99,6 @@ class BookmarkedStories
             }
         }
 
-        return (bool) $isDeleted;
+        return $id;
     }
 }