+++ app/Repos/HackerNews/HnApi.php
@@ -15,21 +15,15 @@ use stdClass;
 
 class HnApi
 {
-    protected $baseUri;
+    protected string $baseUri;
 
-    protected $topStoriesUri;
+    protected string $jobStoriesUri;
 
-    protected $bestStoriesUri;
+    protected string $updatesUri;
 
-    protected $newStoriesUri;
+    protected string $itemUriFormat;
 
-    protected $jobStoriesUri;
-
-    protected $updatesUri;
-
-    protected $itemUriFormat;
-
-    protected $concurrency = 10;
+    protected int $concurrency = 10;
 
     /**
      * Init class properties with common configuration values
@@ -37,9 +31,6 @@ class HnApi
     public function __construct()
     {
         $this->baseUri = config('hackernews.api_base_uri');
-        $this->topStoriesUri = config('hackernews.api_top_stories_uri');
-        $this->bestStoriesUri = config('hackernews.api_best_stories_uri');
-        $this->newStoriesUri = config('hackernews.api_new_stories_uri');
         $this->jobStoriesUri = config('hackernews.api_job_stories_uri');
         $this->updatesUri = config('hackernews.api_updates_uri');
         $this->itemUriFormat =
@@ -114,41 +105,9 @@ class HnApi
     /**
      * @return string
      */
-    public function getTopStoriesUri()
-    {
-        return $this->topStoriesUri;
-    }
-
-    /**
-     * @return string
-     */
-    public function getBestStoriesUri()
-    {
-        return $this->bestStoriesUri;
-    }
-
-    /**
-     * @return string
-     */
-    public function getNewStoriesUri()
-    {
-        return $this->newStoriesUri;
-    }
-
-    /**
-     * @return string
-     */
     public function getJobStoriesUri()
     {
         return $this->jobStoriesUri;
-    }
-
-    /**
-     * @return string
-     */
-    public function getUpdatesUri()
-    {
-        return $this->updatesUri;
     }
 
     /**